+++ library/exit.js
