+++ library/getDataFromCoingecko.js
